added new model. Welcome email and verification of email address

+++ controller/UserController.js
@@ -5,14 +5,17 @@ const bcrypt = require('bcryptjs');
 const jwt = require('jsonwebtoken');
 const {promisify} = require('util');
 const {OAuth2Client} = require("google-auth-library");
-const client = new OAuth2Client(process.env.CLIENT_ID , process.env.CLIENT_SECRET);
 const sendEmail = require("../utils/Email");
+const crypto = require("crypto");
+
+const client = new OAuth2Client(process.env.CLIENT_ID , process.env.CLIENT_SECRET);
 
 const signJWT = async (user_id)=>{
     return await promisify(jwt.sign)(user_id,process.env.JWT_SECRET);
 }
 
 exports.RegisterUser = async (_res, userDetails) => {
+    try {
     const { name, email, phoneNumber, password, securityWord, exams } = userDetails;
 
     let active = false;
@@ -21,45 +24,68 @@ exports.RegisterUser = async (_res, userDetails) => {
     if (!password){
        active = true;
        googleLogin = true
+    } 
+    const userExists = await User.findOne({email: email});
+
+    if (userExists) {
+        throw new ClientError("User already exists!");
     }
+
+    //same for google login and regular email
+    const user = new User({
+        name, 
+        email, 
+        phoneNumber, 
+        password, 
+        securityWord, 
+        exams,
+        active,
+        googleLogin
+    });
     
-    try {
-        const user = new User({
-            name, 
-            email, 
-            phoneNumber, 
-            password, 
-            securityWord, 
-            exams,
-            active,
-            googleLogin
-        });
 
         const salt = await bcrypt.genSalt(10);
         user.securityWord = await bcrypt.hash(securityWord,salt);
 
+        //only for regular signup
         if(password){
             user.password = await bcrypt.hash(password,salt);
-        }
-
-        await user.save();
-        if(password){
             try {
                 const verifyEmail = new EmailVerify({
-                    user : user.id,
+                    name: user.name,
+                    email : user.email, 
+                    phoneNumber : user.phoneNumber, 
+                    password : user.password, 
+                    securityWord : user.securityWord, 
+                    exams: user.exams,
                 });
                 
                 let token = verifyEmail.generateRandomToken();
                 await verifyEmail.save();
-                await sendEmail(token , email);  
+ 
+                const subject = `Verification Email!!`
+                const content = 
+                `
+                <h2>Please verify you email account in order to access the services.</h2>
+                <p>Click on this <a href="http://localhost:4000/users/verifyAccount/${token}">link</a> to verify:</p>
+                <h5>This link shall expire in 1 hour.</h5>
+                `
+
+                await sendEmail(email,subject, content);  
+
                 return;
+                
             } catch (err) {
     
                 await User.findByIdAndDelete(user.id);
                 throw err;
             }
+        } else{
+            await user.save();
         }
 
+        return;
+        
     } catch (err) {
         console.log("error : " , err);
         throw err;
@@ -72,10 +98,7 @@ exports.LoginUser = async (login, res, next) => {
         
         const { email, password, securityWord } = login;
         const user = await User.findOne({ email: email });
-        if (!user.active){
-            return next(new AuthorizationError("Email id not verified!"))
-        }
-    
+
         if(!user) {
             return next(new ClientError("Invalid credentials!"));
         }
@@ -152,4 +175,37 @@ exports.GoogleSignIn = async (body)=>{
        console.log("in the user controller : " , err);
        throw err;
    }
+}
+
+exports.VerifyEmailAccount = async (token)=>{
+    try{
+       
+        let hashedToken = crypto.createHash("SHA256").update(token).digest("hex");
+        let user = await EmailVerify.findOne({token : hashedToken});
+
+        if(!user){
+            throw new AuthorizationError("Invalid Url/token");
+        }
+
+        //storing in the original database
+        await User.create(
+            {
+                name : user.name,
+                phoneNumber : user.phoneNumber,
+                email : user.email,
+                password : user.password,
+                securityWord: user.securityWord,
+                active : true,
+                exams : user.exams        
+            }
+        );
+
+        //deleting from the temporary database
+        await EmailVerify.findByIdAndDelete(user.id);
+        return;
+
+    }catch(err){
+        console.log("errorrr : " , err);
+        throw new err;
+    }
 }
